+++ src/observatory/src/api/http.rs
@@ -1,5 +1,5 @@
-use crate::certificates::http::response::transform_certificate_response as transform_certificate;
 use crate::notifications::http::response::transform_response;
+use crate::openid::http::response::transform_certificate_response as transform_certificate;
 use ic_cdk::management_canister::{HttpRequestResult, TransformArgs};
 use ic_cdk_macros::query;
