+++ src/observatory/src/api/openid.rs
@@ -1,5 +1,5 @@
-use crate::certificates::scheduler::{start_openid_scheduler, stop_openid_scheduler};
 use crate::guards::caller_is_admin_controller;
+use crate::openid::scheduler::{start_openid_scheduler, stop_openid_scheduler};
 use ic_cdk_macros::update;
 
 #[update(guard = "caller_is_admin_controller")]
