+++ src/observatory/src/lib.rs
@@ -1,12 +1,12 @@
 #![deny(clippy::disallowed_methods)]
 
 mod api;
-mod certificates;
 mod console;
 mod guards;
 mod impls;
 mod memory;
 mod notifications;
+mod openid;
 mod random;
 mod store;
 mod templates;
