+++ src/observatory/src/openid/http/request.rs
@@ -1,4 +1,4 @@
-use crate::certificates::http::constants::FETCH_MAX_RESPONSE_BYTES;
+use crate::openid::http::constants::FETCH_MAX_RESPONSE_BYTES;
 use crate::types::state::OpenIdProvider;
 use ic_cdk::management_canister::{
     http_request as http_request_outcall, HttpRequestArgs, TransformContext, TransformFunc,
