+++ src/observatory/src/openid/mod.rs
@@ -1,4 +1,4 @@
+mod certificate;
 mod constants;
 pub mod http;
 pub mod scheduler;
-mod update;
